fastapi-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
numpy
pandas
python-dotenv
google-generativeai